feat(app): moved sync to chain logic to sidebar
